chore(dnsdist): Move various UDP cross-protocol functions and classes to dnsdist::udp
